Move import to parsers
